fix: force fp32 weights in DeBERTa (HF checkpoint is fp16, causes NaN)

+++ scripts/05_train_deberta.py
@@ -38,8 +38,11 @@ def main():
     tok = AutoTokenizer.from_pretrained(args.model)
     # SDPA attention NaNs with DeBERTa's disentangled attention (known
     # transformers bug) - force the eager implementation.
+    # The HF checkpoint is stored in fp16; DeBERTa-v3 attention overflows
+    # in fp16, so we force fp32 weights explicitly.
     model = AutoModelForSequenceClassification.from_pretrained(
-        args.model, num_labels=2, attn_implementation="eager"
+        args.model, num_labels=2, attn_implementation="eager",
+        torch_dtype=torch.float32,
     )
     device = "cuda" if torch.cuda.is_available() else "cpu"
     model.to(device)
